requirements.txt
astropy-healpix
inflection
ligo-gracedb
mocpy
sqlalchemy[postgresql_psycopg2binary]
